requirements.txt
arrow==0.17.0
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.3
biopython==1.78
blessed==1.17.12
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-heroku==0.3.1
django-picklefield==3.0.1
django-q==1.3.4
djangorestframework==3.12.2
gunicorn==20.0.4
h11==0.12.0
httpcore==0.12.2
httpx==0.16.1
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
numpy==1.19.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
rfc3986==1.4.0
service-identity==18.1.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
wcwidth==0.2.5
whitenoise==5.2.0
xmltodict==0.12.0
zope.interface==5.2.0
